# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/authentication":{"get":{"operationId":"AuthenticationController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Authentication"],"security":[{"bearer":[]}]}},"/authentication/login":{"post":{"operationId":"AuthenticationController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/sales":{"post":{"operationId":"SalesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvoiceDto"}}}},"responses":{"201":{"description":""}},"tags":["Sales"],"security":[{"bearer":[]}]},"get":{"operationId":"SalesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Sales"],"security":[{"bearer":[]}]}},"/sales/{id}":{"get":{"operationId":"SalesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Sales"],"security":[{"bearer":[]}]},"put":{"operationId":"SalesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateInvoiceDto"}}}},"responses":{"200":{"description":""}},"tags":["Sales"],"security":[{"bearer":[]}]},"delete":{"operationId":"SalesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Sales"],"security":[{"bearer":[]}]}}},"info":{"title":"API GATEWAY","description":"The API GATEWAY for the application","version":"1.0","contact":{}},"tags":[{"name":"API GATEWAY","description":""}],"servers":[{"url":"api/v1"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"Enter JWT token","in":"header"}},"schemas":{"LoginDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"InvoiceDetailDto":{"type":"object","properties":{"product":{"type":"string"},"quantity":{"type":"number"},"unit_price":{"type":"number"},"invoiceId":{"type":"number"}},"required":["product","quantity","unit_price","invoiceId"]},"InvoiceDto":{"type":"object","properties":{"client":{"type":"string"},"details":{"type":"array","items":{"$ref":"#/components/schemas/InvoiceDetailDto"}}},"required":["client","details"]},"UpdateInvoiceDetailDto":{"type":"object","properties":{"id":{"type":"number"},"product":{"type":"string"},"quantity":{"type":"number"},"unit_price":{"type":"number"}},"required":["id","product","quantity","unit_price"]},"UpdateInvoiceDto":{"type":"object","properties":{"client":{"type":"string"},"details":{"type":"array","items":{"$ref":"#/components/schemas/UpdateInvoiceDetailDto"}}},"required":["client","details"]}}}}